More robust ROI BinVolume creation
Previously this would crash if attempting to create a BinVolume from a
ROI when it included coordinates that were outside the scope of the
referenced image series. Now, this is handled gracefully, with a warning
added if the ROI refers to a slice that is not present in the referenced
image series.