Adds support for DRR generation
This commits adds initial support for computing digitally reconstructed
radiographs (DRRs) in rtkit, by implementing a radiological path
algorithm in ruby (ray.rb), which uses information from the RTPLAN and
CT Image series to create a projection image, which can be converted to
an RTIMAGE DICOM file.

Note that this is just initial, 'alpha level' code, which definitely has
some issues with peformance, functionality and test coverage. Hoewever,
with some polish, this could become quite usable and enable a relatively
easy way of creating custom DRRs based on given RTPLAN/RTSTRUCT/CT
series.

Note also that RTPLAN or RTSTRUCT ROI image delineation/burn-in is not
supported yet. This will have to be added in a future update.