Add support for CR images in RTKIT
These images are loaded in a 'CRSeries' in the 'Image' class. This class
is intended to be a 'general' image class, holding images of modality
'CR' (these film images can be either slice or projection type images,
without it being possible to know at load time).

Note that the CRSeries class is currently unspecced. Possibly the image
series classes will be reorganised at a later time.