Added method for transferring ROI pixels
The new ROI#export_pixels method allows the transfer of a pixel
structure (e.g. table top) from one image set (CT series) to another, by
using the ROI delineation of the structure.

Non-exact pixel locations are matched using a nearest neighbour
algorithm.

This is a basic implementation, which may have some weaknesses to be
improved, but it does seem to work fine based on the included spec
example.