Introduced a new POI class
Previously, POIs were treated as ROIs (which they basically are in the
DICOM standard).

With this new implementation, there is a separation between POIs and
ROIs, where the POI object is much simpler than a ROI, which should make
POI management in RTKIT more enjoyable.

A new Structure class has been introduced which both the POI and ROI
classes inherit from.